requirements.txt

tensorflow>=2.12
streamlit>=1.28
pillow
pandas
reportlab
numpy
